google-cloud-run/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow